refactor people models
