--- backend/src/models/User.ts
@@ -1,35 +1,0 @@
-import mongoose, { Document, Schema } from 'mongoose';
-
-interface IUser extends Document {
-  email: string;
-  name: string;
-  password: string;
-  is_admin: boolean;
-}
-
-const UserSchema = new Schema<IUser>({
-  email: {
-    type: String,
-    required: true,
-    unique: true,
-    trim: true,
-    lowercase: true
-  },
-  name: {
-    type: String,
-    required: true,
-    trim: true
-  },
-  password: {
-    type: String,
-    required: true
-  },
-  is_admin: {
-    type: Boolean,
-    default: false
-  }
-}, {
-  timestamps: true
-});
-
-export default mongoose.model<IUser>('User', UserSchema);
